rename package main -> customers
