+++ pkg/customers/watchman.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an Apache License
 // license that can be found in the LICENSE file.
 
-package main
+package customers
 
 import (
 	"context"
